requirements.txt
# --------- python --------- #
numpy
matplotlib
pandas
scipy
notebook

# --------- pytorch --------- #
torch==2.0.0
torchaudio
torchvision==0.15.0
pytorch-lightning==2.0.1
jsonargparse[signatures]
torchmetrics

# --------- loggers --------- #
# tensorboard
wandb
# neptune-client
# mlflow
# comet-ml

# --------- custom --------- #
av
Pillow
opencv-python
dtw-python
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/pytorchvideo.git

# --------- misc --------- #
tqdm
rich
pyrootutils
python-dotenv
Jinja2
